requirements.txt
# Core FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.6
python-multipart==0.0.6
aiofiles==23.2.1

# AI and news processing
anthropic==0.49.0
langchain-anthropic==0.3.10
langchain-core==0.3.50
feedparser==6.0.11
beautifulsoup4==4.13.3

# Database (ADDED - this was missing!)
sqlalchemy[asyncio]==2.0.23
psycopg[binary]==3.2.10
psycopg-pool==3.2.6

# Caching (simplified to avoid conflicts)
fakeredis==2.30.0

# HTTP and networking
requests==2.32.3
aiohttp==3.9.1

# Configuration
python-dotenv==1.1.0
pydantic==2.11.7

# JSON processing
orjson==3.10.18

# Scheduling
APScheduler==3.11.0

# Stock API dependencies
asyncio-throttle==1.0.2
python-dateutil==2.8.2

# Utilities
packaging==24.2
typing_extensions==4.14.0
certifi==2025.6.15
charset-normalizer==3.4.2
urllib3==2.5.0
idna==3.10
# Add these lines to requirements.txt

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0